Fix a warning from mistyped fakes

+++ tests/Drivers/mh_z19.h
@@ -1,5 +1,7 @@
 
 #pragma once
+#include <stdint.h>
+
 typedef enum
 {
     ser_USART0 = 0,
@@ -18,6 +20,6 @@ typedef enum
     MHZ19_PPM_MUST_BE_GT_999 /**< Span calibration is only allowed from ppm >= 1000 */
 } mh_z19_returnCode_t;
 
-mh_z19_returnCode_t mh_z19_getCo2Ppm(int *ppm);
+mh_z19_returnCode_t mh_z19_getCo2Ppm(uint16_t *ppm);
 void mh_z19_initialise(serial_comPort_t com_port);
 mh_z19_returnCode_t mh_z19_takeMeassuring(void);
