+++ tests/fakes.h
@@ -111,7 +111,7 @@ DECLARE_FAKE_VALUE_FUNC(uint16_t, hih8120_getHumidityPercent_x10);
 
 // ---------------------------- ~ MH_Z19 ~ ---------------------------------------
 
-DECLARE_FAKE_VALUE_FUNC(mh_z19_returnCode_t, mh_z19_getCo2Ppm, int *);
+DECLARE_FAKE_VALUE_FUNC(mh_z19_returnCode_t, mh_z19_getCo2Ppm, uint16_t *);
 
 DECLARE_FAKE_VOID_FUNC(mh_z19_initialise, serial_comPort_t);
